git-svn-id: http://ownageclan.com/svn/war3source@261 e104397d-def3-435e-8314-01917568762b

+++ extensionsource/war3ext/extension.cpp
@@ -392,7 +392,7 @@ static cell_t W3ExtRegister(IPluginContext *pCtx, const cell_t *params)
 {
 	char* strarg1;
 	pCtx->LocalToString(params[1], &strarg1);
-	PRINT("%d smx loaded\n",strarg1);
+	PRINT("%d smx loaded\n",(int)strarg1);
 
 	helpergetfunc=pCtx->GetFunctionByName("Get");
 	if(helpergetfunc==NULL){
